update error message

+++ src/routes/schema/schema.controller.ts
@@ -16,7 +16,7 @@ router.post('/', (req, res) => {
       }
       return schemaService.saveDataset(req.body).then((result) => {
         if (result < 0) {
-          res.status(500).send();
+          res.status(500).send("grahQL error");
         } else {
           return schemaService.deletePriorSavedFields(req.body.id);
         }
@@ -34,9 +34,9 @@ router.post('/', (req, res) => {
          }
         }).then ( (result) => {
           if (result < 0 ) {
-            res.status(500).send();
+            res.status(500).send("graphQL error");
           } else {
-            res.status(200).send();
+            res.status(200).send("dataset saved");
           }
       }).catch ((err) => {res.status(500).send(err.messge); });
 });
@@ -44,12 +44,12 @@ router.get('/user/:userid', (req, res, next) => {
     // console.log (req.params.userid);
     return schemaService.getAllDatasetsOfUser(req.params.userid).then(datasets => {
       if( datasets < 0 ) {
-          return res.status(404).send();
+          return res.status(404).send("resource not found");
       } else {
           return res.status(200).send(datasets)
       }
   }).catch(() => {
-      return res.status(500).send(); //TODO: Introduce better error handling
+      return res.status(500).send("unknown server error."); //TODO: Introduce better error handling
   })
 });
 
@@ -57,32 +57,35 @@ router.get('/dataset/:assetid', (req, res, next) => {
   console.log (req.params.assetid);
   return schemaService.getAdataset(req.params.assetid).then(datasets => {
     if( datasets < 0 ) {
-        return res.status(404).send();
+        return res.status(404).send("resource not found");
     } else {
         return res.status(200).send(datasets)
     }
 }).catch(() => {
-    return res.status(500).send(); //TODO: Introduce better error handling
+    return res.status(500).send("unknown server error."); //TODO: Introduce better error handling
 })
 });
 
 
 router.delete('/dataset/:assetid', (req, res, netxt) => {
     console.log (req.params.assetid);
+    if (req.params.assetid === undefined) {
+      return res.status(404).send("not found)")
+    }
     return schemaService.deleteSchema(req.params.assetid).then(result=> {
       if( result  < 0 ) {
-          return res.status(404).send();
+          return res.status(404).send("not found");
       } else {
           return schemaService.deletePriorSavedFields(req.params.assetid);
       }
   }).then ( (result) => {
         if (result < 0 ) {
-          res.status(500).send();
+          res.status(500).send("graphQL server error.");
         } else {
-          res.status(200).send();
+          res.status(200).send(JSON.stringify(result));
         }
     }).catch(() => {
-    return res.status(500).send(); //TODO: Introduce better error handling
+    return res.status(500).send("unknow server error."); //TODO: Introduce better error handling
   });
 });
